implement authorization using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -7,7 +7,11 @@ use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
-{
+{   
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
 
     public function index()
     {   
@@ -53,12 +57,15 @@ class QuestionsController extends Controller
 
     public function edit(Question $question)
     {
+        $this->authorize("update", $question);
         return view("questions.edit", compact('question'));
     }
 
 
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question);
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your question has been updated');
@@ -66,7 +73,9 @@ class QuestionsController extends Controller
 
 
     public function destroy(Question $question)
-    {
+    {   
+        $this->authorize("delete", $question);
+
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
